separate code-debugger component
