+++ core/gui/src/app/app.module.ts
@@ -139,6 +139,7 @@ import { HubWorkflowComponent } from "./hub/component/workflow/hub-workflow.comp
 import { HubWorkflowSearchBarComponent } from "./hub/component/workflow/search-bar/hub-workflow-search-bar.component";
 import { HubWorkflowDetailComponent } from "./hub/component/workflow/detail/hub-workflow-detail.component";
 import { BreakpointConditionInputComponent } from "./workspace/component/code-editor-dialog/breakpoint-condition-input/breakpoint-condition-input.component";
+import { CodeDebuggerComponent } from "./workspace/component/code-editor-dialog/code-debugger.component";
 
 registerLocaleData(en);
 
@@ -228,6 +229,7 @@ registerLocaleData(en);
     HubWorkflowResultComponent,
     GoogleLoginComponent,
     BreakpointConditionInputComponent,
+    CodeDebuggerComponent
   ],
   imports: [
     BrowserModule,
