+++ core/gui/src/app/workspace/component/code-editor-dialog/code-debugger.component.html
@@ -1,0 +1,7 @@
+<texera-breakpoint-condition-input
+  [operatorId]="this.currentOperatorId"
+  [lineNum]="breakpointConditionLine"
+  [mouseX]="breakpointConditionMouseX"
+  [mouseY]="breakpointConditionMouseY"
+  (closeEmitter)="closeBreakpointConditionInput()"></texera-breakpoint-condition-input>
+
